chatbot/requirements.txt
fastapi==0.128.8
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
uvicorn==0.39.0
httpx==0.27.0
groq==0.9.0